+++ examples/ecsTest.h
@@ -12,45 +12,7 @@
 #include "Engine\ECS\World.h"
 #include "TestFramework.hpp"
 
-//HASH MAP TEST
-#include "Engine/Core/Containers/HopscotchHashMap.h"
-#include "Engine/Core/Containers/SparseHopscotch.h"
-
-//EVENT TEST
-#include "Engine\ECS\Events\EventQueue.hpp"
-#include "Engine\ECS\Events\Signal.hpp"
-
-#include "Engine\Core\JobManager.h"
-
 using namespace ECS::test;
-
-// First let's define the event struct. e is the event type, priority determines the priority.
-struct MyEvent
-{
-	int e;
-	int priority;
-};
-
-// The comparison function object used by eventpp::OrderedQueueList.
-// The function compares the event by priority.
-struct MyCompare
-{
-	template <typename T>
-	bool operator() (const T& a, const T& b) const {
-		return a.template getArgument<0>().priority > b.template getArgument<0>().priority;
-	}
-};
-
-// Define the EventQueue policy
-struct MyPolicy
-{
-	template <typename Item>
-	using QueueList = ECS::EVENT::OrderedQueueList<Item, MyCompare>;
-
-	static int getEvent(const MyEvent& event) {
-		return event.e;
-	}
-};
 
 struct Position
 {
@@ -78,7 +40,7 @@ void hashMapBenchmarks();
 
 int test()
 {
-	RUN_TEST("test_jobSystem",1);
+	RUN_TEST("entityTest",1);
 	//RUN_TEST("test_particalJobSystem", 450);
 
 	//RUN_TEST("test_bigJobSystem",1);
@@ -86,475 +48,6 @@ int test()
 	//RUN_PRIORITY_TESTS(false);
 
 	return 0;
-}
-
-void TestSystemB(ECS::World& world) {
-	std::printf("B is updated\n");
-}
-void TestSystemC(ECS::World& world) {
-	std::printf("C is updated\n");
-}
-void TestSystemD(ECS::World& world) {
-	std::printf("D is updated\n");
-}
-void TestSystemE(ECS::World& world) {
-	std::printf("E is updated\n");
-}
-void TestSystemF(ECS::World& world) {
-	std::printf("F is updated\n");
-}
-
-namespace ECS::System{
-	struct PreInitialization{};
-	struct Initialization{};
-}
-
-TEST_CASE_PRIORITY(test_SystemSchedule) {
-	ECS::World world;
-
-	world = ECS::World();
-
-	world.initialize();
-
-	auto A = world.addSystem<ECS::System::PreInitialization>([](ECS::World& w) {
-		std::printf("A is updated\n");
-		});
-	auto C = world.addSystem<ECS::System::PreInitialization>(TestSystemC);
-	auto B = world.addSystem<ECS::System::PreInitialization>(TestSystemB);
-
-	auto F = world.addSystem<ECS::System::Initialization>(TestSystemF);
-	auto D = world.addSystem<ECS::System::Initialization>(TestSystemD);
-	auto E = world.addSystem<ECS::System::Initialization>(TestSystemE);
-	
-
-	// A → B
-	world.addBefore(A, B); 
-
-	// B → C
-	world.addAfter(B, C);    // F は E の後
-
-	// D → E
-	world.addBefore(D, E);   // F は D より前
-
-	// E → F
-	world.addAfter(E, F);    // C は D の後
-
-	float dt = 0;
-	world.update(dt);
-}
-
-int generateRandomInt(int minNum, int maxNum) {
-	std::random_device rd;
-	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> distribution(minNum, maxNum);
-	return distribution(gen);
-}
-
-void busyWait(std::chrono::nanoseconds duration) {
-	auto start_time = std::chrono::high_resolution_clock::now();
-	while (std::chrono::high_resolution_clock::now() - start_time < duration) {
-		// CPU時間を消費するためのアクティブな待機
-		// ループの中で何もしない
-	}
-}
-
-struct TestHoge
-{
-	int value = 2;
-
-	TestHoge(int v) : value(v){};
-};
-
-struct TestEmptyJob
-{
-
-	TestHoge connecter;
-
-	void execute() {
-		connecter.value = connecter.value * connecter.value;
-		//return connecter.value;
-	}
-};
-
-struct TestPrintJob
-	: public ECS::JobSystem::IJob<TestPrintJob>
-{
-	TestHoge hoge;
-
-	/*static void Execute(TestJob*job) {
-		job->connecter.value = job->connecter.value * job->connecter.value;
-		
-	}*/
-
-	TestPrintJob(int value) : hoge(value){
-	}
-
-	void execute() {
-		hoge.value = hoge.value * hoge.value;
-	}
-};
-
-struct TestJob
-	: public ECS::JobSystem::IJob<TestJob>
-{
-	TestHoge hoge;
-
-	/*static void Execute(TestJob*job) {
-		job->connecter.value = job->connecter.value * job->connecter.value;
-
-	}*/
-
-	TestJob(int value) : hoge(value) {
-	}
-
-	void execute() {
-		hoge.value = hoge.value * hoge.value;
-	}
-};
-
-struct TestParticalJob
-	: public ECS::JobSystem::IJob<TestParticalJob>
-{
-	std::string name;
-
-	/*static void Execute(TestJob*job) {
-		job->connecter.value = job->connecter.value * job->connecter.value;
-
-	}*/
-
-	TestParticalJob(std::string n) : name(n){}
-
-	void execute() {
-		std::printf("%s particalJob executed \n",name.c_str());
-	}
-};
-
-struct TestDelayParticalJob
-	: public ECS::JobSystem::IJob<TestDelayParticalJob>
-{
-	std::string name;
-	int delayTime;
-
-	/*static void Execute(TestJob*job) {
-		job->connecter.value = job->connecter.value * job->connecter.value;
-
-	}*/
-
-	TestDelayParticalJob(std::string n,int delay) : name(n),delayTime(delay) {}
-
-	void execute() {
-		std::printf("%s particalJob executed \n", name.c_str());
-		busyWait(std::chrono::milliseconds(delayTime));
-	}
-};
-
-//connecter.value = connecter.value * connecter.value;
-		//int sleepTime = generateRandomInt(5, 10);
-		//busyWait(std::chrono::milliseconds(sleepTime));
-		/*auto value = this->jobResult.value * this->jobResult.value;
-		this->jobResult.value = value;
-
-		return this->jobResult.value;*/
-		//connecter.value = connecter.value * connecter.value;
-		//return connecter.value;
-
-TEST_CASE_PRIORITY(test_jobSystem) {
-	int check = 9'0000;
-
-	auto recorder = std::make_unique<ECS::JobSystem::TimelineRecorder>();
-	auto& jm = ECS::JobSystem::JobManager::Instance();
-	jm.initialize(9'0000,7,std::move(recorder));
-
-	/*for(int i = 0;i < 20;i++){
-		job->schedule(ECS::JobSy
-		stem::JobCategory::BackGround);
-
-		job.AddRequest(std::make_unique<ECS::JobSystem::FuncCmd<TestJob>>([&check](TestJob& t) {
-			t.connecter.value = check;
-			}));
-	}*/
-
-	//int check = 5625;
-	///本実装では、scheduleIJobでJobIdを発行、dense配列に本体を突っ込む。するようにして、
-	/// schedule時に他のdenseIdを見つつ、denseIdをソートしてchunk化、連続実行できるように作る。
-	/// 
-	///JobHandleに、JobIdを設定して、実際にchunk化で入れられるのはdense側のId
-	//createIJobをなくし、継承先をIJobかどうかで作成は普通にTestJob job();
-
-	//バッチ処理できるようにする。
-	// 今までのtaskArenaのcurrentSlotによるchunk詰めと同じことをする。
-	// JobCategoryにBatch処理用のカテゴリーを追加
-	// check 個のジョブをスケジュール
-	for (int i = 0; i <check; ++i) {
-		auto job = TestJob::create(i);
-
-		auto handle = job->scheduleIJob(ECS::JobSystem::TaskCategory::Batch);
-	}
-
-	std::printf("RealTimeJob Start is %zu\n", jm.getStats().scheduledJobCount());
-
-	auto globalStart = ECS::JobSystem::now();
-	jm.start();
-
-	jm.getStats().waitForAll();
-
-	auto globalEnd = ECS::JobSystem::now();
-
-	std::printf("realTime job %zu finished!! \n", check);
-
-	int  globalDuration = ECS::JobSystem::duration(globalStart, globalEnd);
-	std::cout << "Total duration: "
-		<< globalDuration << " ms\n";
-
-	//ログを取り出してスレッド別タイムラインを出力
-	//auto& dataMap = jm.getRecorder()->getDataMap();
-	//ECS::JobSystem::printTimelines(dataMap, globalStart, globalDuration);
-}
-
-TEST_CASE_PRIORITY(test_chainJobSystem){
-
-	auto recorder = std::make_unique<ECS::JobSystem::TimelineRecorder>();
-	auto& jm = ECS::JobSystem::JobManager::Instance();
-	jm.initialize(100,7, std::move(recorder));
-
-	jm.start();
-
-	//単一依存
-	/*A → B
-	B は A 完了後に実行されることを確認。*/
-	{
-		auto jobA = TestParticalJob::create("A");
-		auto handle = jobA->scheduleIJob();
-
-		auto jobB = TestParticalJob::create("B");
-		auto handle2 = jobB->scheduleIJob(handle);
-	}
-
-	jm.getStats().waitForAll();
-
-	/*複数依存(fan - in)
-		A, B → C*/
-	{
-		std::vector<ECS::JobSystem::JobHandle>handles;
-
-		auto jobA = TestParticalJob::create("A");
-		auto handle = jobA->scheduleIJob();
-
-		auto jobB = TestParticalJob::create("B");
-		auto handle2 = jobB->scheduleIJob();
-
-		handles.push_back(handle);
-		handles.push_back(handle2);
-
-		auto jobC = TestParticalJob::create("C");
-		auto handle3 = jobC->scheduleIJob(handles);
-
-		// 少しずつずらしてスケジューリング
-		busyWait(std::chrono::milliseconds(2));
-	}
-
-	jm.getStats().waitForAll();
-
-	/*複数子依存 (fan-out)
-		A,B → C, D。*/
-	{
-		
-		std::vector<ECS::JobSystem::JobHandle>handles;
-
-		auto jobA = TestDelayParticalJob::create("A",4);
-		auto handle = jobA->scheduleIJob();
-
-		auto jobB = TestDelayParticalJob::create("B",4);
-		auto handle2 = jobB->scheduleIJob();
-
-		handles.push_back(handle);
-		handles.push_back(handle2);
-
-		auto jobC = TestParticalJob::create("C");
-		auto handle3 = jobC->scheduleIJob(handles);
-
-		auto jobD = TestParticalJob::create("D");
-		auto handle4 = jobD->scheduleIJob(handles);
-
-		// 少しずつずらしてスケジューリング
-		busyWait(std::chrono::milliseconds(2));
-	}
-
-	jm.getStats().waitForAll();
-
-	/*深い依存チェーン
-		A → B → C → D
-		順序通りに実行されることを確認。*/
-	{
-		auto jobA = TestParticalJob::create("A");
-		auto handle = jobA->scheduleIJob();
-
-		auto jobB = TestDelayParticalJob::create("B",2);
-		auto handle2 = jobB->scheduleIJob(handle);
-
-		auto jobC = TestParticalJob::create("C");
-		auto handle3 = jobC->scheduleIJob(handle2);
-
-		auto jobD = TestParticalJob::create("D");
-		auto handle4 = jobD->scheduleIJob(handle3);
-
-		// 少しずつずらしてスケジューリング
-		busyWait(std::chrono::milliseconds(2));
-	}
-
-	jm.getStats().waitForAll();
-
-	//assertTrue(jm.checkRanAllJobInJobQueues(), "JobSystem Valid Test");
-	//assertTrue(!jm.isAbort(), "JobSystem Work Test");
-	//ECS::JobSystem::printTimelines(dataMap, globalStart, globalDuration);
-}
-
-//struct TestConnector
-//{
-//	std::vector<int> resultData;
-//
-//	TestConnector() = default;
-//};
-//
-struct TestParallelJob
-	: public ECS::JobSystem::IParallelJob<TestParallelJob>
-{
-	std::vector<int>results;
-
-	TestParallelJob(size_t size) : results(size,0){};
-
-	inline void Execute(size_t index) {
-		//results[index] = index * index;
-		std::printf("%zu is result is %d \n",index, results[index]);
-	}
-
-	
-};
-
-struct TestPrintParallelJob
-	: public ECS::JobSystem::IParallelJob<TestPrintParallelJob>
-{
-	std::vector<int>results;
-	std::string name;
-
-	TestPrintParallelJob(size_t size,std::string n) : results(size, 0),name(n) {};
-
-	inline void Execute(size_t index) {
-		std::printf("%s TestPrintParallelJob[%zu] Executed \n",name.c_str(),index);
-	}
-
-	~TestPrintParallelJob() {
-		std::printf("%s deleted \n", name.c_str());
-	}
-};
-
-//パラレルジョブテスト
-TEST_CASE_ORDER(test_bigJobSystem) {
-	auto recorder = std::make_unique<ECS::JobSystem::TimelineRecorder>();
-	auto& jm = ECS::JobSystem::JobManager::Instance();
-	jm.initialize(100,7, std::move(recorder));
-
-	/*size_t resultSize = 100;
-
-	size_t batchSize = 10;
-
-	size_t workerCount = 2;*/
-
-	size_t resultSize = 5;
-
-	size_t batchSize = 1;
-
-	size_t workerCount = 2;
-
-	/*size_t resultSize = 10'0000;
-
-	size_t batchSize = 1000;
-
-	size_t workerCount = 7;*/
-	{
-		auto parallelJob = TestPrintParallelJob::create(resultSize, "A");
-
-		auto handle = parallelJob->schedule(resultSize, batchSize, workerCount);
-	}
-
-	std::printf("ParallelJob %zu Start is %d\n", jm.getStats().scheduledJobCount(),resultSize/batchSize);
-
-	auto globalStart = ECS::JobSystem::now();
-
-	jm.start();
-
-	jm.getStats().waitForAll();
-
-	// 5) テスト終了時刻を記録＆全体持続時間を計算
-	auto globalEnd = ECS::JobSystem::now();
-	int  globalDuration = ECS::JobSystem::duration(globalStart, globalEnd);
-	std::cout << "Total duration: "
-		<< globalDuration << " ms\n";
-}
-
-//パラレルジョブに依存関係追加
-TEST_CASE_ORDER(test_chainBigJobSystem) {
-	auto recorder = std::make_unique<ECS::JobSystem::TimelineRecorder>();
-	auto& jm = ECS::JobSystem::JobManager::Instance();
-	jm.initialize(100, 7, std::move(recorder));
-
-	size_t resultSize = 5;
-
-	size_t batchSize = 1;
-
-	size_t workerCount = 2;
-
-	jm.start();
-
-	/*複数依存(fan - in)
-		A → B*/
-	{
-		auto testIJob = TestDelayParticalJob::create("A", 2);
-
-		auto handle = testIJob->scheduleIJob();
-
-		auto parallelJob = TestPrintParallelJob::create(resultSize, "B");
-
-		parallelJob->schedule(resultSize, batchSize, workerCount, handle);
-	}
-
-	jm.getStats().waitForAll();
-
-	/*複数依存(fan - in)
-		A, B → C*/
-	{
-		auto testIJob = TestDelayParticalJob::create("A", 2);
-
-		auto handle = testIJob->scheduleIJob();
-
-		auto testIJob2 = TestDelayParticalJob::create("B", 4);
-
-		auto handle2 = testIJob2->scheduleIJob();
-
-		std::vector<ECS::JobSystem::JobHandle>handles;
-		handles.push_back(handle);
-		handles.push_back(handle2);
-
-		auto parallelJob = TestPrintParallelJob::create(resultSize,"C");
-
-		parallelJob->schedule(resultSize, batchSize, workerCount, handles);
-	}
-
-	jm.getStats().waitForAll();
-
-	/*複数依存(fan - in)
-		A → B*/
-	{
-		auto parallelJob = TestPrintParallelJob::create(resultSize, "A");
-
-		auto handle = parallelJob->schedule(resultSize, batchSize, workerCount);
-
-		auto parallelJob2 = TestPrintParallelJob::create(resultSize,"B");
-
-		parallelJob2->schedule(resultSize, batchSize, workerCount, handle);
-	}
-
-	jm.getStats().waitForAll();
 }
 
 TEST_CASE(test_sort_empty) {
@@ -619,669 +112,7 @@ TEST_CASE_PRIORITY(test_sort_with_custom_algo) {
 	ASSERT(std::is_sorted(v.begin(), v.end(),std::less<>{}), "custom algo sorts");
 }
 
-bool myCallback(const int a, const std::unique_ptr<int>& b) {
-	return a < *b;
-}
 
-bool myFallback() {
-	return false;
-}
-
-inline ECS::ecs_map::id_type hash_idApple() { return 1; }
-inline ECS::ecs_map::id_type hash_idBanana() { return 2; }
-
-//SignalEventDispatherのテスト
-TEST_CASE(world_event_test)
-{
-	struct Apple{
-		int x = 5;
-		int y = 0;
-	};
-
-	struct Banana {
-		int x = 30;
-		int y = 0;
-		Banana(int x_, int y_) : x(x_), y(y_) {}
-		Banana() = default;
-	};
-
-	struct Grape {
-		int x = 30;
-		int y = 0;
-	};
-
-	ECS::EVENT::Signal bus;                     // EventType = int
-	
-	// 読み取り専用のListener
-	bus.connect([](const Apple* a) {
-		std::cout << "const Apple x = " << a->x << "\n";
-		});
-	
-	// 変更可能なListener
-	bus.connect([](Apple* a) {
-		a->x += 10;
-		std::cout << "Apple*a->x = " << a->x << "\n";
-		});
-
-	// shared_ptrも素のポインタと同様に渡せる
-	bus.connect([](const std::unique_ptr<const Banana>* s) {
-		std::cout << "const std::unique_ptr<Banana>* s->x  = " << s->get()->x << "\n";
-		});
-	
-	bus.connect([](std::unique_ptr<Banana>* s) {
-		std::cout << "std::unique_ptr<Banana>& s->x  = " << s->get()->x << "\n";
-		s->get()->x = 100;
-		});
-	
-	bus.connect([](const std::shared_ptr<Grape>* s) {
-		std::cout << "const std::shared_ptr<Grape>&s->x  = " << s->get()->x << "\n";
-		});
-
-	bus.connect([](std::shared_ptr<Grape>* s) {
-		std::cout << "std::shared_ptr<Grape>&->x  = " << s->get()->x << "\n";
-		s->get()->x += 50;
-		});
-	
-	// 発行側：ポインタ or スマートポインタだけ書けば OK
-	Apple a{ 42 };
-	const Apple ca{ 99 };
-	
-	const std::unique_ptr<const Banana> cb = std::make_unique<Banana>(10,10);
-	std::unique_ptr<Banana> b = std::make_unique<Banana>();
-	std::shared_ptr<Grape> g = std::make_shared<Grape>();
-	const std::shared_ptr<Grape> cg = std::make_shared<Grape>();
-
-	bus.publish(&a);
-	bus.publish(&ca);
-	bus.publish(&cb);
-	bus.publish(&b);
-	bus.publish(&cg);
-	bus.publish(&g);
-
-	//Banana b{ 7 };
-	//bus.publish(&b);   // Banana* リスナ呼び出し
-
-	// フレーム末に一括ディスパッチ
-	bus.dispatch();
-
-	std::cout << "change unique_ptr<Banana>b->x = " << b->x << std::endl;
-	std::cout << "change shared_ptr<Grapge>g->x = " << g->x << std::endl;
-
-	//bus.dispatch();
-
-	//std::cout << "banana y = " << b.y << '\n';   // 10
-
-	// 発行後も a/b は依然 発行側が所有
-	
-}
-
-//Signalのwaitイベントテスト
-TEST_CASE(test_wait_unblocks_on_event){
-	std::atomic<bool> unblocked{ false };
-	struct Apple {
-		int x = 5;
-		int y = 0;
-	};
-
-	struct Banana {
-		int x = 30;
-		int y = 0;
-	};
-
-	ECS::EVENT::Signal bus;
-
-	// 1) ワーカーはまず wait() でブロック
-	std::thread worker([&] {
-		bus.wait();
-		unblocked = true;
-		});
-
-	// 2) 少し待ってブロッキング中であることを確認
-	std::this_thread::sleep_for(std::chrono::milliseconds(100));
-	assert(unblocked == false);
-
-	// 3) イベントを publish & dispatch して wakeup させる
-	Apple a{ 777 };
-	bus.publish(&a);
-	bus.dispatch();
-
-	// 4) worker が目覚めるまで少し待機
-	std::this_thread::sleep_for(std::chrono::milliseconds(100));
-	assert(unblocked == true);
-
-	worker.join();
-	std::cout << "test_wait_unblocks_on_event: OK\n";
-}
-
-TEST_CASE(test_waitFor_times_out_and_succeeds){
-	using namespace std::chrono;
-	struct Apple {
-		int x = 5;
-		int y = 0;
-	};
-
-	struct Banana {
-		int x = 30;
-		int y = 0;
-	};
-	Apple a{ 123 };
-	ECS::EVENT::Signal bus;
-
-	// 1) 空のままで waitFor がタイムアウトすること
-	auto t0 = high_resolution_clock::now();
-	bool got = false;
-	got = bus.waitFor(milliseconds(200));
-	auto dt = duration_cast<milliseconds>(high_resolution_clock::now() - t0);
-	assert(got == false);
-	assert(dt >= milliseconds(190));         // ほぼ200ms待った
-
-	std::cout << "waitFor timeout: OK ( waited "
-		<< dt.count() << " ms )\n";
-
-	// 2) イベント到着後ならすぐ戻ること
-	//    先に publish しておく
-	bus.publish(&a);
-
-	t0 = high_resolution_clock::now();
-	got = bus.waitFor(milliseconds(500));
-	dt = duration_cast<milliseconds>(high_resolution_clock::now() - t0);
-	assert(got == true);
-	assert(dt < milliseconds(50));           // ほぼ即戻り
-
-	// dispatch を呼ぶのを忘れずに
-	bus.dispatch();
-
-	std::cout << "waitFor wakeup: OK ( waited "
-		<< dt.count() << " ms )\n";
-}
-
-TEST_CASE(testEventQueue)
-{
-	
-	{
-		ECS::EVENT::EventQueueBase<int, void(const std::string&, std::unique_ptr<int>&)> queue;
-
-		queue.appendListener(3, [](const std::string& s, std::unique_ptr<int>& n) {
-			std::cout << "Got event 3, s is " << s << " n is " << *n << std::endl;
-			});
-		// The listener prototype doesn't need to be exactly same as the dispatcher.
-		// It would be find as long as the arguments is compatible with the dispatcher.
-		queue.appendListener(5, [](std::string s, const std::unique_ptr<int>& n) {
-			std::cout << "Got event 5, s is " << s << " n is " << *n << std::endl;
-			});
-		queue.appendListener(5, [](const std::string& s, std::unique_ptr<int>& n) {
-			std::cout << "Got another event 5, s is " << s << " n is " << *n << std::endl;
-			});
-
-		// Enqueue the events, the first argument is always the event type.
-		// The listeners are not triggered during enqueue.
-		queue.enqueue(3, "Hello", std::unique_ptr<int>(new int(38)));
-		queue.enqueue(5, "World", std::unique_ptr<int>(new int(58)));
-
-		queue.processOne();
-	}
-	
-
-	{
-		ECS::EVENT::EventQueueBase<int, void(const int&, std::unique_ptr<int>&)> queue;
-
-		queue.appendListener(3, [](const int s, std::unique_ptr<int>& n) {
-			std::cout << "Got event 3, s is " << s << " n is " << *n << std::endl;
-			});
-		// The listener prototype doesn't need to be exactly same as the dispatcher.
-		// It would be find as long as the arguments is compatible with the dispatcher.
-		queue.appendListener(5, [](const int s, const std::unique_ptr<int>& n) {
-			std::cout << "Got event 5, s is " << s << " n is " << *n << std::endl;
-			});
-
-		// Enqueue the events, the first argument is always the event type.
-		// The listeners are not triggered during enqueue.
-		queue.enqueue(3, 2, std::unique_ptr<int>(new int(5)));
-		queue.enqueue(5, 12, std::unique_ptr<int>(new int(3)));
-		
-		queue.processIf([](const int a, const std::unique_ptr<int>& b){
-			return a > *b;
-		});
-		
-		queue.processIf(myCallback);
-
-		queue.processIf(myFallback);
-
-		queue.processUntil([]() {
-			return false;
-			});
-	}
-	
-	
-	{
-	ECS::EVENT::EventQueueBase<int, void(const std::string&, std::unique_ptr<int>&)> queue;
-
-	queue.appendListener(3, [](const std::string& s, std::unique_ptr<int>& n) {
-		std::cout << "Got event 3, s is " << s << " n is " << *n << std::endl;
-		});
-	// The listener prototype doesn't need to be exactly same as the dispatcher.
-	// It would be find as long as the arguments is compatible with the dispatcher.
-	queue.appendListener(5, [](std::string s, const std::unique_ptr<int>& n) {
-		std::cout << "Got event 5, s is " << s << " n is " << *n << std::endl;
-		});
-	queue.appendListener(5, [](const std::string& s, std::unique_ptr<int>& n) {
-		std::cout << "Got another event 5, s is " << s << " n is " << *n << std::endl;
-		});
-
-	// Enqueue the events, the first argument is always the event type.
-	// The listeners are not triggered during enqueue.
-	queue.enqueue(3, "Hello", std::unique_ptr<int>(new int(38)));
-	queue.enqueue(5, "World", std::unique_ptr<int>(new int(58)));
-
-	// Process the event queue, dispatch all queued events.
-	queue.process();
-
-	std::cout << std::endl;
-	}
-	
-	///////////////2
-	
-	{
-	using EQ = ECS::EVENT::EventQueueBase<int, void(int)>;
-	EQ queue2;
-
-	constexpr int stopEvent = 1;
-	constexpr int otherEvent = 2;
-	
-	// Start a thread to process the event queue.
-	// All listeners are invoked in that thread.
-	std::thread thread([stopEvent, otherEvent, &queue2]() {
-		volatile bool shouldStop = false;
-		queue2.appendListener(stopEvent, [&shouldStop](int) {
-			shouldStop = true;
-			});
-		queue2.appendListener(otherEvent, [](const int index) {
-			std::cout << "Got event, index is " << index << std::endl;
-			});
-		
-		while (!shouldStop) {
-			queue2.wait();
-
-			queue2.process();
-		};
-		
-	});
-	
-	// Enqueue an event from the main thread. After sleeping for 10 milliseconds,
-	// the event should have be processed by the other thread.
-	queue2.enqueue(otherEvent, 1);
-	std::this_thread::sleep_for(std::chrono::milliseconds(10));
-	std::cout << "Should have triggered event with index = 1" << std::endl;
-
-	queue2.enqueue(otherEvent, 2);
-	std::this_thread::sleep_for(std::chrono::milliseconds(10));
-	std::cout << "Should have triggered event with index = 2" << std::endl;
-
-	{
-		// EventQueueBase::DisableQueueNotify is a RAII class that
-		// disables waking up any waiting threads.
-		// So no events should be triggered in this code block.
-		// DisableQueueNotify is useful when adding lots of events at the same time
-		// and only want to wake up the waiting threads after all events are added.
-		EQ::DisableQueueNotify disableNotify(&queue2);
-
-		queue2.enqueue(otherEvent, 10);
-		std::this_thread::sleep_for(std::chrono::milliseconds(10));
-		std::cout << "Should NOT trigger event with index = 10" << std::endl;
-
-		queue2.enqueue(otherEvent, 11);
-		std::this_thread::sleep_for(std::chrono::milliseconds(10));
-		std::cout << "Should NOT trigger event with index = 11" << std::endl;
-	}
-	// The DisableQueueNotify object is destroyed here, and has resumed
-	// waking up waiting threads.
-	//So the events should be triggered.
-	std::this_thread::sleep_for(std::chrono::milliseconds(10));
-	std::cout << "Should have triggered events with index = 10 and 11" << std::endl;
-
-	queue2.enqueue(stopEvent, 1);
-	thread.join();
-	}
-	
-	{
-	std::cout << std::endl << "EventQueue tutorial 3, ordered queue" << std::endl;
-
-	using EQ = ECS::EVENT::EventQueueBase<int, void(const MyEvent&), MyPolicy>;
-	EQ queue3;
-
-	queue3.appendListener(3, [](const MyEvent& event) {
-		std::cout << "Get event " << event.e << "(should be 3)."
-			<< " priority: " << event.priority << std::endl;
-		});
-	queue3.appendListener(5, [](const MyEvent& event) {
-		std::cout << "Get event " << event.e << "(should be 5)."
-			<< " priority: " << event.priority << std::endl;
-		});
-	queue3.appendListener(7, [](const MyEvent& event) {
-		std::cout << "Get event " << event.e << "(should be 7)."
-			<< " priority: " << event.priority << std::endl;
-		});
-
-	// Add an event, the first number 5 is the event type, the second number 100 is the priority.
-	// After the queue processes, the events will be processed from higher priority to lower priority.
-	
-	queue3.enqueue(MyEvent{ 5, 100 });
-	queue3.enqueue(MyEvent{ 5, 200 });
-	queue3.enqueue(MyEvent{ 7, 300 });
-	queue3.enqueue(MyEvent{ 7, 400 });
-	queue3.enqueue(MyEvent{ 3, 500 });
-	queue3.enqueue(MyEvent{ 3, 600 });
-	
-	queue3.process();
-	}
-	
-	{
-
-		// This is the definition of event types
-		enum class EventType
-		{
-			// for MouseEvent
-			mouse,
-
-			// for KeyboardEvent
-			keyboard,
-
-			// for either MouseEvent or KeyboardEvent, it's used to demonstrate
-			// how the listener detects event type dynamically
-			input,
-
-			// for ChangedEvent
-			changed
-		};
-
-		// This is the base event. It has a getType function to return the actual event type.
-		class Event
-		{
-		public:
-			explicit Event(const EventType type) : type(type) {
-			}
-
-			virtual ~Event() {
-			}
-
-			EventType getType() const {
-				return type;
-			}
-
-		private:
-			EventType type;
-		};
-
-		class MouseEvent : public Event
-		{
-		public:
-			MouseEvent(const int x, const int y)
-				: Event(EventType::mouse), x(x), y(y)
-			{
-			}
-
-			int getX() const { return x; }
-			int getY() const { return y; }
-
-		private:
-			int x;
-			int y;
-		};
-
-		class KeyboardEvent : public Event
-		{
-		public:
-			explicit KeyboardEvent(const int key)
-				: Event(EventType::keyboard), key(key)
-			{
-			}
-
-			int getKey() const { return key; }
-
-		private:
-			int key;
-		};
-
-		class ChangedEvent : public Event
-		{
-		public:
-			explicit ChangedEvent(const std::string& text)
-				: Event(EventType::changed), text(text)
-			{
-			}
-
-			std::string getText() const { return text; }
-
-		private:
-			std::string text;
-		};
-
-		// We will pass event as EventPointer, here it's std::shared_ptr<Event>.
-		// It allows EventQueue to store the events in internal buffer without slicing the objects
-		// in asynchronous API (EventQueue::enqueue and EventQueue::process, etc).
-		// If we only use the synchronous API (EventDispatcher, or EventQueue::dispatch),
-		// we can dispatch events as reference.
-		using EventPointer = std::shared_ptr<Event>;
-
-		// We are going to dispatch event objects directly without specify the event type explicitly,
-		// so we need to define policy to let eventpp know how to get the event type from the event object.
-		struct EventPolicy
-		{
-			static EventType getEvent(const EventPointer& event) {
-				return event->getType();
-			}
-		};
-
-		std::cout << std::endl << "EventQueue tutorial 4, typical event system in an application" << std::endl;
-		
-		using EQ = ECS::EVENT::EventQueueBase<EventType, void(const EventPointer&), EventPolicy>;
-		EQ queue;
-
-		queue.appendListener(EventType::mouse, [](const EventPointer& event) {
-			const MouseEvent* mouseEvent = static_cast<const MouseEvent*>(event.get());
-			std::cout << "Received mouse event, x=" << mouseEvent->getX() << " y=" << mouseEvent->getY()
-				<< std::endl;
-			});
-		queue.appendListener(EventType::keyboard, [](const EventPointer& event) {
-			const KeyboardEvent* keyboardEvent = static_cast<const KeyboardEvent*>(event.get());
-			std::cout << "Received keyboard event, key=" << (char)keyboardEvent->getKey()
-				<< std::endl;
-			});
-		queue.appendListener(EventType::input, [](const EventPointer& event) {
-			std::cout << "Received input event, ";
-			if (event->getType() == EventType::mouse) {
-				const MouseEvent* mouseEvent = static_cast<const MouseEvent*>(event.get());
-				std::cout << "it's mouse event, x=" << mouseEvent->getX() << " y=" << mouseEvent->getY()
-					<< std::endl;
-			}
-			else if (event->getType() == EventType::keyboard) {
-				const KeyboardEvent* keyboardEvent = static_cast<const KeyboardEvent*>(event.get());
-				std::cout << "it's keyboard event, key=" << (char)keyboardEvent->getKey() << std::endl;
-			}
-			else {
-				std::cout << "it's an event that I don't understand." << std::endl;
-			}
-			});
-		queue.appendListener(EventType::changed, [](const EventPointer& event) {
-			const ChangedEvent* changedEvent = static_cast<const ChangedEvent*>(event.get());
-			std::cout << "Received changed event, text=" << changedEvent->getText() << std::endl;
-			});
-
-		// Asynchronous API, put events in to the event queue.
-		queue.enqueue(std::make_shared<MouseEvent>(123, 567));
-		queue.enqueue(std::make_shared<KeyboardEvent>('W'));
-		queue.enqueue(std::make_shared<ChangedEvent>("This is new text"));
-		queue.enqueue(EventType::input, std::make_shared<MouseEvent>(10, 20));
-		// then process all events.
-		queue.process();
-
-		// Synchronous API, dispatch events to the listeners directly.
-		queue.dispatch(std::make_shared<KeyboardEvent>('Q'));
-		queue.dispatch(EventType::input, std::make_shared<ChangedEvent>("Should not display"));
-	}
-	
-}
-
-TEST_CASE(testDispatch){
-	// The namespace is eventpp
-// The first template parameter int is the event type,
-// the event type can be any type such as std::string, int, etc.
-// The second is the prototype of the listener.
-	ECS::EVENT::EventDispatcher_Multi<int, void()> dispatcher;
-
-	// Add a listener. As the type of dispatcher,
-	// here 3 and 5 is the event type,
-	// []() {} is the listener.
-	// Lambda is not required, any function or std::function
-	// or whatever function object with the required prototype is fine.
-	dispatcher.appendListener(3, []() {
-		std::cout << "Got event 3." << std::endl;
-		});
-
-	auto removeHandle = dispatcher.appendFrontListener(3, []() {
-		std::cout << "Got first event 3." << std::endl;
-		});
-
-	dispatcher.appendListener(5, []() {
-		std::cout << "Got event 5." << std::endl;
-		});
-	auto handle = dispatcher.appendListener(5, []() {
-		std::cout << "Got third event 5." << std::endl;
-		});
-
-	dispatcher.insertListener(5, []() {
-		std::cout << "Got second event 5." << std::endl;
-		},handle);
-	
-	// Dispatch the events, the first argument is always the event type.
-	dispatcher.dispatch(3);
-	dispatcher.dispatch(5);
-
-	dispatcher.removeListener(3, removeHandle);
-	dispatcher.removeListener(3, removeHandle);
-	dispatcher.dispatch(3);
-
-	assertTrue(dispatcher.hasAnyListener(3) == true, "dispatcher.hasAnyListener(3) == true");
-
-	assertTrue(dispatcher.haveHandle(5,handle) == true, "dispatcher.haveHandle(5,handle) == true");
-
-	std::cout << std::endl;
-	
-	// The listener has two parameters.
-	ECS::EVENT::EventDispatcher_Multi<int, void(const std::string&, const bool)> dispatcher2;
-
-	dispatcher2.appendListener(3, [](const std::string& s, const bool b) {
-		std::cout << std::boolalpha << "Got event 3, s is " << s << " b is " << b << std::endl;
-		});
-	// The listener prototype doesn't need to be exactly same as the dispatcher.
-	// It would be find as long as the arguments is compatible with the dispatcher.
-	dispatcher2.appendListener(5, [](std::string s, int b) {
-		std::cout << std::boolalpha << "Got event 5, s is " << s << " b is " << b << std::endl;
-		});
-	dispatcher2.appendListener(5, [](const std::string& s, const bool b) {
-		std::cout << std::boolalpha << "Got another event 5, s is " << s << " b is " << b << std::endl;
-		});
-
-	// Dispatch the events, the first argument is always the event type.
-	dispatcher2.dispatch(3, "Hello", true);
-	dispatcher2.dispatch(5, "World", false);
-
-	// Define an Event to hold all parameters.
-	struct MyEvent {
-		std::string type;
-		std::string message;
-		int param;
-	};
-
-	// Define policies to let the dispatcher knows how to
-	// extract the event type.
-	struct MyEventPolicies
-	{
-		static std::string getEvent(const MyEvent& e, bool ) {
-			return e.type;
-		}
-	};
-
-	ECS::EVENT::EventDispatcher_Multi<
-		std::string,
-		void(const MyEvent&, bool),
-		MyEventPolicies
-	>dispatcher3;
-
-	// Add a listener.
-// Note: the first argument is the event type of type int, not MyEvent.
-
-	dispatcher3.appendListener("a", [](const MyEvent& e, bool b) {
-		std::cout
-			<< std::boolalpha
-			<< "Got event 3" << std::endl
-			<< "Event::type is " << e.type << std::endl
-			<< "Event::message is " << e.message << std::endl
-			<< "Event::param is " << e.param << std::endl
-			<< "b is " << b << std::endl
-			;
-		});
-
-	// Dispatch the event.
-	// The first argument is Event.
-	dispatcher3.dispatch(MyEvent{ "a","Hello world", 38 }, true);
-	
-}
-
-std::string testCallbackFunction(const bool& flag, const std::string& message){
-	return flag ? "Callback 3: " + message : "Callback 3: Condition not met.";
-}
-
-struct testCallbackFunctionClass {
-	std::string memberFunction(const bool& flag, const std::string& message){
-		return flag ? "Callback 4: " + message : "Callback 4: Condition not met.";
-	}
-};
-
-// テスト関数
-TEST_CASE(testCallbackList) {
-
-	using CL = ECS::EVENT::CallbackList_Multi<std::string(const bool&, const std::string&)>;
-	CL callbackList;
-
-	auto handle1 = callbackList.append([](const bool& flag, const std::string& message) -> std::string {
-		return flag ? "Callback 1: " + message : "Callback 1: Condition not met.";
-		});
-
-	auto handle2 = callbackList.append([](const bool& flag, const std::string& message) -> std::string {
-		return flag ? "Callback 2: " + message : "Callback 2: Condition not met.";
-		});
-
-	auto handle3 = callbackList.append<&testCallbackFunction>();
-
-	testCallbackFunctionClass testClass;
-
-	auto handle4 = callbackList.append<&testCallbackFunctionClass::memberFunction>(testClass);
-
-	std::string result1 = handle1.lock()->callback(true, "Hello Handle");
-	assertTrue(result1 == "Callback 1: Hello Handle", "Callback 1 returned expected value.");
-
-	std::string result2 = handle2.lock()->callback(false, "Hello Handle");
-	assertTrue(result2 == "Callback 2: Condition not met.", "Callback 2 returned expected value.");
-
-	std::string result3 = handle3.lock()->callback(false, "Hello Handle");
-	assertTrue(result3 == "Callback 3: Condition not met.", "Callback 3 returned expected value.");
-
-	std::string result4 = handle4.lock()->callback(false, "Hello Handle");
-	assertTrue(result4 == "Callback 4: Condition not met.", "Callback 4 returned expected value.");
-
-	callbackList.remove(handle2);
-	std::cout << "\n--- After Removal Callback2---\n";
-	callbackList.foreach([](const CL::CallbackType& callback) {
-		std::string result = callback(true, "Hello Callback");
-		std::cout<< result << std::endl;
-	});
-
-	auto results = callbackList(true,"Hello CallbackVector");
-	for(auto&x:results){
-		std::cout<<x<<std::endl;
-	}
-}
 
 TEST_CASE(typeListTest)
 {
@@ -1293,110 +124,86 @@ TEST_CASE(typeListTest)
 	static_assert(ECS::type_Index_v<float, MyList> != static_cast<std::size_t>(ECS::npos), "float is not in the list");
 }
 
-// --- テスト関数 ---
-//<Key,Value>
-//benchmark(IndexBase_HopscotchHashMap<int,int>(capacity : 8u), "IndexBase");
-//benchmark(SparseSetBase_HopscotchHashMap<int, int> sparseSetMap(cap : 100000), "SparseSet");
-/*
-結果の分析(Debug)
-##挿入時間 (IndexBase: 18ms → SparseSet: 15ms)
-- SparseSetとほぼ同じレベルまで短縮され、バケット探索のオーバーヘッドが削減されたことがわかる。
-- Hopscotchの局所性を活かしつつ、リハッシュの影響を抑えられた ことが成功の要因かも。
-##検索時間 (IndexBase: 4ms / SparseSet: 5ms)
-- ほぼ差がなく、どちらの方式も高速な検索ができている。
-- これは適切なハッシュ計算とバケット配置のおかげ。
-##イテレーション時間 (IndexBase: 4ms → SparseSet: 12ms)
-- SparseSet方式は dense にデータを密集させるために余計なループ処理が発生しやすいかも。
-- 一方、IndexBase方式は next() による空バケットスキップが効率的に機能している。
-
-ankerl::unordered_dense::map - Insert Time: 71 ms
-ankerl::unordered_dense::map - Find Time: 47 ms [100000 found]
-ankerl::unordered_dense::map - Iteration Time: 11 ms [1409965408 sum]
-
-結果の分析(Release)
-##挿入時間 (Insert Time)
-- IndexBase: 452ms
-- SparseSet: 157ms
-SparseSet方式のほうが約3倍高速にデータを挿入 できています。
-これは、SparseSetが 連続したメモリアクセスを活用し、キャッシュミスを最小限に抑える ためです。
-IndexBase方式では ホップ情報 (hopinfoes_) の管理やバケット移動 (moveEmpty()) が挿入のコストを増やしている 可能性があります。
-##検索 (Find Time)
-- IndexBase: 237ms
-- SparseSet: 279ms
-IndexBase方式のほうが若干検索が速い（SparseSetより約42ms短縮）。
-これは、IndexBase方式が ホップ情報を活用して、局所性の高い検索ができるため です。
-一方、SparseSetは 探索するテーブルが小さいので、検索自体は比較的軽いが、メモリアクセスがやや分散している可能性がある。
-##イテレーション (Iteration Time)
-- IndexBase: 9ms
-- SparseSet: 2ms
-SparseSet方式のほうが約4倍高速にイテレーションできている。
-SparseSetは dense にデータを格納するため、シーケンシャルアクセスが最適化される のに対し、
-IndexBase方式では バケットをスキップしながらイテレーションするため、オーバーヘッドが発生している可能性 があります。
-*/
-
-template<typename HashMapType>
-void benchmark(HashMapType&, const std::string&, const int);
-
-void hashMapBenchmarks()
-{
-	struct Health {};
-	struct OwnerA {};
-	struct OwnerB {};
-	struct OwnerC {};
-
-	ECS::ecs_map::HopscotchHashMap<int, int> indexBaseMap(10000000);
-	//ecs_map::SparseHopscotchHashMap<int,int> sparseSetMap(10000);
-	//ecs_map::unordered_dense_map<int,int> unordered_denseMap(10000);
-
-	std::cout << "Benchmarking...\n";
-	benchmark(indexBaseMap, "IndexBase", 10000000);
-
-	//benchmark(sparseSetMap, "SparseSet", 10000);
-}
-
-#pragma optimize("", off)  // 最適化を無効化
-template<typename HashMapType>
-void benchmark(HashMapType& map, const std::string& name, const int numElements) {
-	using namespace std::chrono;
-
-	// 挿入テスト
-	auto start = high_resolution_clock::now();
-	for (int i = 0; i < numElements; ++i) {
-		map.insert(i, i * 2);
-	}
-	auto end = high_resolution_clock::now();
-	std::cout << name << " - Insert Time: " << duration_cast<milliseconds>(end - start).count() << " ms\n";	
-
-	// 検索テスト
-	start = high_resolution_clock::now();
-	int found = 0;
-	for (int i = 0; i < numElements; ++i) {
-		if (map.find(i)){
-			found++;
-		}
-	}
-
-	end = high_resolution_clock::now();
-	std::cout << name << " - Find Time: " << duration_cast<milliseconds>(end - start).count() << " ms ["<< found<<" found"<< "]\n";
-
-	// イテレーションテスト
-	start = high_resolution_clock::now();
-	volatile long long sum = 0;
-	for (auto it = map.begin(); it != map.end(); ++it) {
-		sum += *it;
-	}
-	end = high_resolution_clock::now();
-	std::cout << name << " - Iteration Time: " << duration_cast<milliseconds>(end - start).count() << " ms [" << sum <<" sum" << "]\n";
-}
-#pragma optimize("", on)  // 最適化をオンに戻す
-
 struct testBasicStorageComponent {
 	static constexpr ECS::COMPONENT::StorageType storage_pref = ECS::COMPONENT::StorageType::BasicType;
 };
 
+TEST_CASE(entityTest)
+{
+	ECS::World world = ECS::World();
+
+	world.initialize();
+
+	Position position = Position(0.0f, 5.0f);
+	auto entity = world.spawn("", position, Velocity(5.0f, 0.1f));
+	auto entity2 = world.spawnEmpty();
+
+	//auto entity2 = ECS::sWorld.spawnEmpty();
+	world.emplaceOrUpdateComponent<Velocity>(entity2, 1.0f, 0.5f);
+	world.emplaceOrUpdateComponent<Position>(entity2);
+	//auto comp = ECS::world().getComponent<Position>(entity);
+
+	//auto view2 = view.Exclude<Position>();
+	//auto view = ECS::world().View<Velocity>(exclude_t<Position>{});
+
+	/*
+	for (size_t i = 0; i < packed.size(); i++) {
+		Position a;
+		std::tie(a) = packed[i].components;
+	}
+	*/
+	auto view = world.View<Position, Velocity>();
+	//auto view2 = view->Exclude<Position>();
+
+	for (auto& x : *view)
+	{
+		auto& entityID = x.entity;
+		auto& vel = view->get<Velocity>(x.components);
+		auto& posi = view->get<Position>(x.components);
+		bool hasComp = world.has<Velocity>(view->get<EntityID>(x));
+	}
+
+	/*
+	for (EntityID x : *view2)
+	{
+		bool hasComp = ECS::world().has<Velocity>(x);
+
+	}
+	*/
+
+	for (auto [entityID, position, velocity] : view->each()) {
+		auto id = entityID;
+		auto posi = position;
+		auto vel = velocity;
+	}
+
+	view->each([](auto entity, auto& pos, auto& vel) {
+		pos.x += 5.0f;
+		vel.x += 5.0f;
+		});
+
+	view->each([](auto& pos, auto& vel) {
+		pos.x += 5.0f;
+		vel.x += 5.0f;
+		});
+
+	//auto entities = scene.getWorld().findEntitiesWithComponents<Velocity>();
+	//auto comp = scene.getWorld().getComponent<Position>(entity);
+	//scene.getWorld().removeComponent<Position>(entity);
+
+	//comp->x+= 10.0f;
+	//comp = scene.getWorld().getComponent<Position>(entity);
+
+	//scene.getWorld().despawn(entity);
+	//scene.getWorld().despawn(entity);
+	MESSAGE("view test Clear");
+}
+
 //作成テスト
 TEST_CASE(test_create_group) {
-	ECS::World world;
+	ECS::World world = ECS::World();
+
+	world.initialize();
 
 	struct A {int x = 0;};
 	struct B {int x = 10;};
@@ -1416,8 +223,10 @@ TEST_CASE(test_create_group) {
 		isPrintDespawnLog = true;
 	};
 
+	//entiyスポーン時にイベント追加(ログ)
 	world.entityPoolConstructor().append(spawnLog);
 	world.entityPoolDestructor().append(destroyLog);
+
 	auto emptyEntity = world.spawnEmpty();
 	world.despawn(emptyEntity);
 
@@ -1539,7 +348,7 @@ TEST_CASE(test_create_group) {
 	std::cout<<"sort\n";
 
 	for(auto&x:sortGroup){
-		//std::cout<<GetEntityIndex(x)<<",";
+		std::cout<<GetEntityIndex(x)<<",";
 	}
 
 	std::cout << "\n";
@@ -1695,112 +504,12 @@ TEST_CASE_DISABLED(testGroupIdentifiers) {
 	//std::cout << "すべての識別 ID のテストが正常に完了しました！" << std::endl;
 }
 
-TEST_CASE(entityTest)
-{
-	ECS::World world;
-
-	Position position = Position(0.0f, 5.0f);
-	auto entity = world.spawn("", position, Velocity(5.0f, 0.1f));
-	auto entity2 = world.spawnEmpty();
-
-	//auto entity2 = ECS::sWorld.spawnEmpty();
-	world.emplaceOrUpdateComponent<Velocity>(entity2, 1.0f, 0.5f);
-	world.emplaceOrUpdateComponent<Position>(entity2);
-	//auto comp = ECS::world().getComponent<Position>(entity);
-
-	//auto view2 = view.Exclude<Position>();
-	//auto view = ECS::world().View<Velocity>(exclude_t<Position>{});
-
-	/*
-	for (size_t i = 0; i < packed.size(); i++) {
-		Position a;
-		std::tie(a) = packed[i].components;
-	}
-	*/
-	auto view = world.View<Position, Velocity>();
-	//auto view2 = view->Exclude<Position>();
-
-	for (auto& x : *view)
-	{
-		auto& entityID = x.entity;
-		auto& vel = view->get<Velocity>(x.components);
-		auto& posi = view->get<Position>(x.components);
-		//bool hasComp = ECS::world().has<Velocity>(view->get<EntityID>(x));
-	}
-
-	/*
-	for (EntityID x : *view2)
-	{
-		bool hasComp = ECS::world().has<Velocity>(x);
-
-	}
-	*/
-
-	for (auto [entityID, position, velocity] : view->each()) {
-		auto id = entityID;
-		auto posi = position;
-		auto vel = velocity;
-	}
-
-	view->each([](auto entity, auto& pos, auto& vel) {
-		pos.x += 5.0f;
-		vel.x += 5.0f;
-		});
-
-	view->each([](auto& pos, auto& vel) {
-		pos.x += 5.0f;
-		vel.x += 5.0f;
-		});
-
-	//auto entities = scene.getWorld().findEntitiesWithComponents<Velocity>();
-	//auto comp = scene.getWorld().getComponent<Position>(entity);
-	//scene.getWorld().removeComponent<Position>(entity);
-
-	//comp->x+= 10.0f;
-	//comp = scene.getWorld().getComponent<Position>(entity);
-
-	//scene.getWorld().despawn(entity);
-	//scene.getWorld().despawn(entity);
-	MESSAGE("view test Clear");
-}
-
-TEST_CASE(test_hopscotchHashMap){
-	//InsertTest
-	{
-		ECS::ecs_map::HopscotchHashMap<EntityIndex, std::string> map;
-		map.insert(1, "hello");
-		auto result = map.find(1);
-		assertTrue("hello"!=*result,"hello!=result"); // ない場合は空文字と比較
-	}
-
-	//find nullTest
-	{
-		ECS::ecs_map::HopscotchHashMap<int, std::string> map;
-		assertTrue(nullptr!=map.find(99),"nullptr!=map.find(99)"); // 存在しないキーは null を返す
-	}
-
-	//eraseTest
-	{
-		ECS::ecs_map::HopscotchHashMap<int, std::string> map;
-		map.insert(1, "hello");
-		map.erase(1);
-		assertTrue(nullptr!=map.find(1),"nullptr!=map.find(1)"); // 削除後は null
-	}
-
-	//clearTest
-	{
-		ECS::ecs_map::HopscotchHashMap<int, std::string> map;
-		 map.insert(1, "hello");
-		 map.insert(2, "world");
-		 assertTrue(2!=map.size(),"2!=map.size()");
-		 map.clear();
-		 assertTrue(0!=map.size(),"0!=map.size()");
-	}
-}
 
 bool executeTimeTest()
 {
 	auto world = ECS::World();
+	world.initialize();
+
 	auto start_creation = std::chrono::high_resolution_clock::now();
 
 	for (size_t i = 0; i < 1000000; i++)
@@ -1897,81 +606,7 @@ bool executeTimeTest_NormalComponentBase() {
 
 //hash関係計測
 
-// 計測用
-using namespace std::chrono;
 
-// FNV-1a ハッシュ
-constexpr uint64_t hash_FNV1a(std::string_view sv) {
-	constexpr uint64_t FNV_OFFSET_BASIS = 0xcbf29ce484222325;
-	constexpr uint64_t FNV_PRIME = 0x100000001b3;
-
-	uint64_t hash = FNV_OFFSET_BASIS;
-	for (char c : sv) {
-		hash ^= static_cast<uint64_t>(c);
-		hash *= FNV_PRIME;
-	}
-	return hash;
-}
-
-// ローリングハッシュ (基数: 31, MOD: 1e9+7)
-constexpr uint64_t hash_Rolling(std::string_view sv) {
-	constexpr uint64_t BASE = 31;
-	constexpr uint64_t MOD = 1000000007;
-
-	uint64_t hash = 0;
-	uint64_t power = 1;
-
-	for (char c : sv) {
-		hash = (hash * BASE + static_cast<uint64_t>(c)) % MOD;
-		power = (power * BASE) % MOD;
-	}
-	return hash;
-}
-
-std::string random_string(size_t length) {
-	static std::mt19937_64 rng(std::random_device{}());
-	static std::uniform_int_distribution<int> dist(97, 122);  // 'a' ~ 'z'
-
-	std::string s;
-	s.reserve(length);
-	for (size_t i = 0; i < length; ++i) {
-		s += static_cast<char>(dist(rng));  // キャストを追加
-	}
-	return s;
-}
-
-void hashTimetest() {
-	constexpr size_t TEST_CASES = 100000;  // テスト回数
-	constexpr size_t STRING_LENGTH = 50;   // 文字列長
-
-	std::vector<std::string> test_strings;
-	for (size_t i = 0; i < TEST_CASES; ++i) {
-		test_strings.push_back(random_string(STRING_LENGTH));
-	}
-
-	// FNV-1a ハッシュの測定
-	auto start_FNV1a = high_resolution_clock::now();
-	for (const auto& str : test_strings) {
-		volatile uint64_t hash = hash_FNV1a(str);
-	}
-	auto end_FNV1a = high_resolution_clock::now();
-	auto duration_FNV1a = duration_cast<microseconds>(end_FNV1a - start_FNV1a).count();
-
-	// ローリングハッシュの測定
-	auto start_Rolling = high_resolution_clock::now();
-	for (const auto& str : test_strings) {
-		volatile uint64_t hash = hash_Rolling(str);
-	}
-	auto end_Rolling = high_resolution_clock::now();
-	auto duration_Rolling = duration_cast<microseconds>(end_Rolling - start_Rolling).count();
-
-	// 結果出力
-	std::cout << "FNV-1a Hash Time: " << duration_FNV1a << " s\n";
-	std::cout << "Rolling Hash Time: " << duration_Rolling << " s\n";
-
-	//FNV - 1a: 17, 445
-	//Rolling Hash : 62, 311
-}
