--- include/Engine/ECS/System/PreInitializationGroup.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "SystemGroup.h"
-
-namespace ECS::System {
-    class PreInitializationGroup : public SystemGroup {
-
-    public:
-        PreInitializationGroup(ECS::World& world) {
-        }
-    };
-}
